just stole all code from the 0.92.x branch from inkscape, and added my code to it, probably forgot a lot of files etc,,

+++ src/libavoid/connector.cpp
@@ -885,7 +885,7 @@ bool ConnRef::generatePath(void)
             {
                 // Check for consecutive points on opposite 
                 // corners of two touching shapes.
-                COLA_ASSERT(abs((long)i->pathNext->id.objID - (long)i->id.objID) != 2);
+                COLA_ASSERT(abs((long)i->pathNext->id.objID - i->id.objID) != 2);
             }
         }
     }
